+++ external/mtools/file_name.h
@@ -31,19 +31,19 @@ struct dos_name_t {
   char sentinel;
 };
 
-int dos_to_wchar(doscp_t *fromDos, const char *dos, wchar_t *wchar, size_t len);
+size_t dos_to_wchar(doscp_t *fromDos, const char *dos, wchar_t *wchar, size_t len);
 void wchar_to_dos(doscp_t *toDos, wchar_t *wchar, char *dos, size_t len, int *mangled);
 
-doscp_t *cp_open(int codepage);
+doscp_t *cp_open(unsigned int codepage);
 void cp_close(doscp_t *cp);
 
-int wchar_to_native(const wchar_t *wchar, char *native,
-		    size_t len, size_t out_len);
+size_t wchar_to_native(const wchar_t *wchar, char *native,
+		       size_t len, size_t out_len);
 
 #define WCHAR_TO_NATIVE(wchar,native,len) \
 	wchar_to_native((wchar),(native),(len),sizeof(native))
 
-int native_to_wchar(const char *native, wchar_t *wchar, size_t len,
-		    const char *end, int *mangled);
+size_t native_to_wchar(const char *native, wchar_t *wchar, size_t len,
+		       const char *end, int *mangled);
 
 #endif
